Time: 298 ms (59.1%), Space: 108.9 MB (31.78%) - LeetHub

+++ 2094-remove-stones-to-minimize-the-total/2094-remove-stones-to-minimize-the-total.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int minStoneSum(vector<int>& piles, int k) {
-         priority_queue<long long> pq;
+         priority_queue<int> pq;
         for(auto it:piles) pq.push(it);
         while(k>0){
             k--;
@@ -10,7 +10,7 @@ public:
             n=n-n/2;
             pq.push(n);
         }
-        long long ans=0;
+        int ans=0;
         while(!pq.empty()){
             ans+=pq.top();
             pq.pop();
